+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When using `embassy` option, `async_main.rs` file was renamed to `main.rs` (#93)
 - The UI no longer allows selecting options with missing requirements, and does not allow deselecting
   options that are required by other options. (#101)
+- Options can now declare negative requirements (e.g. `!alloc` can not be enabled if `alloc` is used) (#101)
 
 ### Fixed